requirements.txt
langchain==0.1.16
langchain-core==0.1.43
langchain-community==0.0.30
openai==1.12.0
python-dotenv==1.0.1
pypdf==4.2.0
streamlit==1.30.0
